requirements.txt
# Project requirements
fastapi
uvicorn
chromadb
ollama
python-dotenv
pytest
